+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(SpaDES.docs)
+
+test_check("SpaDES.docs")